2/3/2014

+++ pc3.cpp
@@ -164,7 +164,7 @@ int rockScissorPaper (char playerOne, char playerTwo)
  */
 int asciiValue (char c)
 {
-	return c;
+	return static_cast<int>(c);
 }
 
 /*
